moved errorAlert to common components
