+++ frontend/src/components/common/ErrorAlert.tsx
